+++ dflib-csv/src/main/java/org/dflib/csv/Csv.java
@@ -1,10 +1,13 @@
 package org.dflib.csv;
 
 import org.dflib.DataFrame;
+import org.dflib.connector.ByteSource;
+import org.dflib.connector.ByteSources;
 
 import java.io.File;
 import java.io.Reader;
 import java.nio.file.Path;
+import java.util.Map;
 
 public class Csv {
 
@@ -19,6 +22,20 @@ public class Csv {
 
     public static DataFrame load(String filePath) {
         return loader().load(filePath);
+    }
+
+    /**
+     * @since 1.1.0
+     */
+    public static DataFrame load(ByteSource src) {
+        return loader().load(src);
+    }
+
+    /**
+     * @since 1.1.0
+     */
+    public static Map<String, DataFrame> loadAll(ByteSources src) {
+        return loader().loadAll(src);
     }
 
     public static DataFrame load(Reader reader) {