+++ dflib-parquet/src/main/java/org/dflib/parquet/Parquet.java
@@ -2,10 +2,12 @@ package org.dflib.parquet;
 
 import org.apache.parquet.schema.MessageType;
 import org.dflib.DataFrame;
+import org.dflib.connector.ByteSource;
+import org.dflib.connector.ByteSources;
 
 import java.io.File;
 import java.nio.file.Path;
-
+import java.util.Map;
 
 
 public class Parquet {
@@ -20,6 +22,20 @@ public class Parquet {
 
     public static DataFrame load(String filePath) {
         return loader().load(filePath);
+    }
+
+    /**
+     * @since 1.1.0
+     */
+    public static DataFrame load(ByteSource src) {
+        return loader().load(src);
+    }
+
+    /**
+     * @since 1.1.0
+     */
+    public static Map<String, DataFrame> loadAll(ByteSources src) {
+        return loader().loadAll(src);
     }
 
     public static MessageType loadSchema(File file) {