+++ RELEASE-NOTES.md
@@ -34,6 +34,7 @@
 * #370 ECharts: support "itemStyle", "symbolSize", "lineStyle" for "line" series
 * #371 ECharts: support "itemStyle" for "boxplot" series
 * #382 $bool(..).cumSum()
+* #413 ByteSource: generic data source for loaders
 * #415 CsvLoader to support an explicit encoding
 * #418 Exception printing null columns